refactor(runtime-providers): extract the OpenCode config target reader from the local provider connector
OpenCodeLocalProviderConnector owned three private methods that resolve which
OpenCode config file a request targets: readConfigTarget, readGlobalConfig and
readProjectConfig, plus the OpenCodeConfigTarget shape they return. That logic
is about the OpenCode config layout on disk, not about the connector's provider
protocol, and it already leaned on four symbols exported from the sibling module
openCodeGlobalConfigOverrides (OPENCODE_GLOBAL_CONFIG_FILENAMES,
OPENCODE_PROJECT_CONFIG_CANDIDATES, assertOpenCodeInlineConfigOverrideSafe and
getOpenCodeGlobalConfigOverrideConflict).

The three methods move into that sibling as readOpenCodeConfigTarget (exported),
readOpenCodeGlobalConfigTarget and readOpenCodeProjectConfigTarget (both kept
module-private, as the class methods were), and OpenCodeConfigTarget moves with
them and becomes exported. No new file is added: the config-layout knowledge now
lives in one place.

Two things are deliberately preserved rather than tidied:

- readOpenCodeConfigTarget is NOT declared async. The original readConfigTarget
  calls assertOpenCodeInlineConfigOverrideSafe synchronously and then returns a
  promise out of a ternary, so an inline-override violation throws synchronously
  instead of arriving as a rejection. Marking the function async would change
  that.
- getEnvironment is passed in as a function port, not as an already-resolved
  environment. createOpenCodeGlobalConfigEnvironmentResolver returns a
  non-memoised closure, and the original code called it twice per read (once in
  readConfigTarget, once again in readGlobalConfig). Passing the resolver keeps
  both calls.

The connector keeps its two call sites, now passing homePath and getEnvironment
explicitly; ensureGlobalDirectory still defaults to false and only writeConfigNow
opts into true. It drops the node:fs import, isPathInside and the
RuntimeLocalProviderScopeDto type import, all of which were used only by the
moved code, and goes from 1018 to 830 lines, staying above the 800-line floor and
188 below its frozen cap. openCodeGlobalConfigOverrides goes from 179 to 389.

Behaviour is unchanged and both connector test suites pass without modification.
Note for reviewers: turning class methods into module-level functions shifts
every moved line two columns left, and prettier then re-joins a few wrapped
expressions, so git will not detect this as a rename even though the bodies are
otherwise identical.
